(refactor) added app directory which is imported for both Android and IOS
