+++ index.android.js
@@ -1,90 +1,4 @@
-'use strict';
-import React, { Component } from 'react';
-import {
-  AppRegistry,
-  Dimensions,
-  StyleSheet,
-  Text,
-  TouchableHighlight,
-  View
-} from 'react-native';
-import Camera from 'react-native-camera';
-var CryptoJS = require('crypto-js');
-var Cloudinary = require('./config/cloudinary.js')
-
-class ComputerVisionMobile extends Component {
-  
-  takePicture() {
-    var context = this;
-    this.camera.capture()
-      .then((data) => {
-        console.log('picture taken')
-        console.log(data)
-        context.uploadImage(data.path);
-      })
-      .catch(err => console.error(err));
-  }
-
-  uploadImage(uri) {
-    console.log('uploadImage called')
-    // DEV NOTE: On future iterations, we can upload without an API Key (and without hashing it with our secret)
-    // But in order to do that, we'll need to set up a Cloudinary PRESET for us to upload to.
-    let timestamp = (Date.now() / 1000 | 0).toString();
-    let api_key = Cloudinary['api_key']
-    let api_secret = Cloudinary['api_secret']
-    let cloud = Cloudinary['cloud_name']
-    let hash_string = 'timestamp=' + timestamp + api_secret
-    let signature = CryptoJS.SHA1(hash_string).toString();
-    let upload_url = 'https://api.cloudinary.com/v1_1/' + cloud + '/image/upload'
-
-    let xhr = new XMLHttpRequest();
-    xhr.open('POST', upload_url);
-    xhr.onload = () => {
-      console.log(xhr);
-    };
-    let formdata = new FormData();
-    formdata.append('file', {uri: uri, type: 'image/jpg', name: 'upload.jpg'});
-    formdata.append('timestamp', timestamp);
-    formdata.append('api_key', api_key);
-    formdata.append('signature', signature);
-    xhr.send(formdata);
-  }
-
-  render() {
-    return (
-      <View style={styles.container}>
-        <Camera
-          ref={(cam) => {
-            this.camera = cam;
-          }}
-          style={styles.preview}
-          aspect={Camera.constants.Aspect.fill}>
-          <Text style={styles.capture} onPress={this.takePicture.bind(this)}>[CAPTURE]</Text>
-        </Camera>
-      </View>
-    );
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1
-  },
-  preview: {
-    flex: 1,
-    justifyContent: 'flex-end',
-    alignItems: 'center',
-    height: Dimensions.get('window').height,
-    width: Dimensions.get('window').width
-  },
-  capture: {
-    flex: 0,
-    backgroundColor: '#fff',
-    borderRadius: 5,
-    color: '#000',
-    padding: 10,
-    margin: 40
-  }
-});
+import  { AppRegistry } from 'react-native';
+import ComputerVisionMobile from './app/index.js';
 
 AppRegistry.registerComponent('ComputerVisionMobile', () => ComputerVisionMobile);
